+++ parallel-consumer-core/src/test/java/io/confluent/csid/actors/ActorImplTest.java
@@ -22,18 +22,18 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 
 /**
- * @see Actor
+ * @see ActorImpl
  */
 @Slf4j
-class ActorTest {
+class ActorImplTest {
 
     public static final String MESSAGE = "tell";
 
     Greeter greeter = new Greeter();
 
-    Actor<Greeter> actor = new Actor<>(greeter);
+    ActorImpl<Greeter> actor = new ActorImpl<>(greeter);
 
-    public ActorTest() {
+    public ActorImplTest() {
         actor.start();
     }
